==> 2-CleanDataset.ipynb <==
# Clean Dataset
This notebook takes all the downloaded data from Coursera and exracts just the course transcripts.

import os
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import re

## Input Course URL
The course id can be found in the hyperlink for any page in the course.

url = 'https://www.coursera.org/learn/siads697698/lecture/3vwIb/how-to-do-a-standup'
course = re.search('(?<=coursera.org/learn/)(\w+)', url).group(0)
lecture = url.split('/')[-1]

print(course)
print(lecture)

## Load Directory

directory = os.listdir('/Users/nicolascap/MADS/Capstone/intermediate_data')
new = True
for file in directory:
    if course in file:
        print("Course Already In Directory")
        new = False

## Create List of Filepaths 

if new:
    directory = os.getcwd()
    filepaths=[]

    for subdir, dirs, files in os.walk(directory):
        for file in files:
            filepath = subdir + os.sep + file

            if filepath.endswith(".txt"):
                c = re.search('(?<=data/)(\w+)', filepath).group(0)
                if c == course:
                    filepaths.append(filepath)
    print(len(filepaths))
    print(filepaths[:5])
else:
    print("No need to run cell, files already in directory")

## Create Transcript DataFrame

if new:
    tup =[]
    for file_path in filepaths: 
        f = open(file_path,'r')
        A = f.read().replace('\n',' ')
        course_id = course
        video_title = file_path.split('/')[-1].split('.')[0][3:]
        if len(A) < 40000: #split long transcripts into two lines 
            tup.append((course_id, video_title, A))
        else: 
            tup.append((course_id, video_title, A[:len(A)//2]))
            tup.append((course_id, video_title, A[len(A)//2:]))

    df = pd.DataFrame(tup,columns=['course_id','video_title', 'transcripts'])
    df['length'] = df['transcripts'].str.len()
    df = df[df.length != 0] 
    df
else:
    print("No need to run cell, files already in directory")

## Save Dataset
We save the transcript dataset as a csv file for further analysis.

if new:
    df.to_csv("./intermediate_data/transcripts_{}.csv".format(course))
else:
    print("No need to run cell, files already in directory")

## Next step
After you saved the dataset here, run the next step in the workflow [3-TranscriptSummarization.ipynb](./3-TranscriptSummarization.ipynb) or go back to [0-Workflow.ipynb](./0-Workflow.ipynb).

---

**Authors:** [Wei Zhou](mailto:weiwzhou@umich.edu), [Nick Capaldini](mailto:nickcaps@umich.edu), University of Michigan, August 21, 2022

---
