==> 3-TranscriptSummarization.ipynb <==
# Transcript Summarization
This notebook reads the transcript dataset and generates summaries using the Bert Extractive Summarizer.

There are 2 methods of Bert Summarization
* Bert Extractive Summarizer
* SBert Summarizer

Hyperparameters to be considered for models
* param ratio: Ratio of sentences to use.
* param min_length: Minimum length of sentence candidates to utilize for the summary.
* param max_length: Maximum length of sentence candidates to utilize for the summary.
* param use_first: Whether or not to use the first sentence.
* param algorithm: Which clustering algorithm to use. (kmeans, gmm)
* param num_sentences: Number of sentences to use (overrides ratio).
* param return_as_list: Whether or not to return sentences as list.

import pandas as pd
import numpy as np
from summarizer import Summarizer
from summarizer.sbert import SBertSummarizer
import re
import os

## Input Course URL
The course id can be found in the hyperlink for any page in the course.

url = 'https://www.coursera.org/learn/siads697698/lecture/3vwIb/how-to-do-a-standup'
course = re.search('(?<=coursera.org/learn/)(\w+)', url).group(0)

## Load Directory

directory = os.listdir('/Users/nicolascap/MADS/Capstone/intermediate_data')
new = True
for file in directory:
    if '{}_summaries'.format(course) in file:
        print("Course Already In Directory")
        new = False
        break

## Read in Transcript DataFrame

df = pd.read_csv("./intermediate_data/transcripts_{}.csv".format(course), index_col=0)
df.head()

## Generate First Transcript to Summarize


#Input url
print(course)
title = url.split('/')[-1]
print(title)

body = df[df['video_title']==title].transcripts.iloc[0]
print(len(body.split('. ')))
print(body)

## Generate Summary

### Initialize Summarizer

model = Summarizer('distilbert-base-uncased', hidden=[-2], hidden_concat=True, random_state=42)

### Generate Result

result = model(body, num_sentences=6, use_first=False) #min_length=60

### Print Result

result

## Apply to All Transcripts


df['summary'] = df['transcripts'].apply(lambda x: model(x, num_sentences=6, use_first=False))

df

## Save DataFrame with Feature Vectors
We save the dataset with protein sequence, fold classification, and feature vectors as a Pandas dataframe for further analysis.

df.to_csv("./intermediate_data/transcripts_{}_summaries.csv".format(course))

## Next step
After you saved the dataset here, run the next step in the workflow [4-KeywordExtraction.ipynb](./4-KeywordExtraction.ipynb) or go back go back to [0-Workflow.ipynb](./0-Workflow.ipynb).

---

**Authors:** [Wei Zhou](mailto:weiwzhou@umich.edu), [Nick Capaldini](mailto:nickcaps@umich.edu), University of Michigan, August 21, 2022

---
