==> 4-KeywordExtraction.ipynb <==
# Keyword Extraction
This notebook reads the transcript dataset and generates keywords using the keyphrase_vectorizers, spacy, and keybert libraries

import pandas as pd
from keyphrase_vectorizers import KeyphraseCountVectorizer
import spacy
from keybert import KeyBERT
from collections import Counter
from nltk.stem import WordNetLemmatizer
import re
import os

## Input Course URL
The course id can be found in the hyperlink for any page in the course.

url = 'https://www.coursera.org/learn/siads697698/lecture/3vwIb/how-to-do-a-standup'
course = re.search('(?<=coursera.org/learn/)(\w+)', url).group(0)

## Load Directory

directory = os.listdir('/Users/nicolascap/MADS/Capstone/intermediate_data')
new = True
for file in directory:
    if '{}_summaries_keywords'.format(course) in file:
        print("Course Already In Directory")
        new = False
        break

## Read in Transcript DataFrame

df = pd.read_csv("./intermediate_data/transcripts_{}_summaries.csv".format(course)).drop(['Unnamed: 0'], axis=1)
df

## Generate First Transcript for Keywords

new_line_full = df.transcripts.iloc[6] 
new_line_summary = df.summary.iloc[6] 

## Initialize Vectorizer

vectorizer = KeyphraseCountVectorizer()
nlp=spacy.load('en_core_web_sm')

## Initialize Model

kw_model = KeyBERT()

## Generate Full Text Keywords

res_full = kw_model.extract_keywords(docs=new_line_full,vectorizer=KeyphraseCountVectorizer(), top_n=10)
res_full

## Generate Summary Text Keywords

res_summary = kw_model.extract_keywords(docs=new_line_summary,vectorizer=KeyphraseCountVectorizer(), top_n=10)
res_summary

## Combine Keywords and Take Top 5

top_5 = res_full + res_summary
top_5.sort(key = lambda x: x[1], reverse=True)
d = dict(top_5)#dict(top_5[:3])
k = [k for k,v in d.items()]
z = [z.split() for z in k]
z = sum(z, [])
lemmatizer = WordNetLemmatizer()
z_1 = [lemmatizer.lemmatize(z_1) for z_1 in z ]

c = Counter(z_1)
countDict = dict(c) 
rep_words = [word for word, occurrences in countDict.items() if occurrences >= 2]

tracker_dict = {word: False for word in rep_words}
remove_list = []
for i in k:
    remove = False
    for word in rep_words:
        if word in i:
            if tracker_dict[word] == False:
                tracker_dict[word] = True
            else:
                #print(word, i)
                remove=True
    if remove:
        remove_list.append(i)

for i in remove_list:
    k.remove(i)
top_5 = k[:5]

top_5

## Apply to All Transcripts

def keyword_creation(transcript, summary):
    res_full = kw_model.extract_keywords(docs=transcript,vectorizer=KeyphraseCountVectorizer(), top_n=10)
    res_summary = kw_model.extract_keywords(docs=summary,vectorizer=KeyphraseCountVectorizer(), top_n=10)
    top_5 = res_full + res_summary
    top_5.sort(key = lambda x: x[1], reverse=True)
    d = dict(top_5)
    k = [k for k,v in d.items()]
    ## simplification
    
    z = [z.split() for z in k]
    z = sum(z, [])
    lemmatizer = WordNetLemmatizer()
    z_1 = [lemmatizer.lemmatize(z_1) for z_1 in z ]

    c = Counter(z_1)
    countDict = dict(c) 
    rep_words = [word for word, occurrences in countDict.items() if occurrences >= 2]

    tracker_dict = {word: False for word in rep_words}
    remove_list = []
    for i in k:
        remove = False
        for word in rep_words:
            if word in i:
                if tracker_dict[word] == False:
                    tracker_dict[word] = True
                else:
                    #print(word, i)
                    remove=True
        if remove:
            remove_list.append(i)

    for i in remove_list:
        k.remove(i)
    top_5 = k[:5]
    ## simplification
    
    return top_5



df['keywords'] = df.apply(lambda x: keyword_creation(x.transcripts, x.summary), axis=1)

df.head()

## Save Dataset
We save the transcript dataset as a csv file for further analysis.

df.to_csv("./intermediate_data/transcripts_{}_summaries_keywords.csv".format(course))

## Next step
After you saved the dataset here, run the next step in the workflow [5-HyperlinkGeneration.ipynb](./5-HyperlinkGeneration.ipynb) or go back to [0-Workflow.ipynb](./0-Workflow.ipynb).

---

**Authors:** [Wei Zhou](mailto:weiwzhou@umich.edu), [Nick Capaldini](mailto:nickcaps@umich.edu), University of Michigan, August 21, 2022

---
