==> 5-HyperlinkGeneration.ipynb <==
# Hyperlink Generation
This notebook generates hyperlinks for the keywords identified in the transcript dataset using the wikipedia library.

import pandas as pd
import wikipedia
import re
import os
import spacy

## Input Course URL
The course id can be found in the hyperlink for any page in the course.

url = 'https://www.coursera.org/learn/siads697698/lecture/3vwIb/how-to-do-a-standup'
course = re.search('(?<=coursera.org/learn/)(\w+)', url).group(0)

course

## Load Directory

directory = os.listdir('/Users/nicolascap/MADS/Capstone/intermediate_data')
new = True
for file in directory:
    if '{}_summaries_keywords_hyperlinks'.format(course) in file:
        print("Course Already In Directory")
        new = False

## Read in Transcript DataFrame

df = pd.read_csv("./intermediate_data/transcripts_{}_summaries_keywords.csv".format(course)).drop(['Unnamed: 0'], axis=1)
df.head()

## Generate First Transcript for URLs

keywords = df.keywords.iloc[10] #7

summary = df.summary.iloc[10] #7
data_science = wikipedia.summary('Data_science')
online_learning_in_higher_education = wikipedia.summary('Online_learning_in_higher_education')

def clean_alt_list(list_):
    list_ = list_.replace("', '", 'X')
    list_ = list_.replace("['", 'X')
    list_ = list_.replace("']", 'X')
    return list_.split("X")[1:-1]
keywords = clean_alt_list(keywords)

## Define get_options Function

def get_options(keyword):
    """All keywords input will get a list of options output"""
    #if type(keyword) != list
    
    options = list(keyword)
    
    
    try:
        page = wikipedia.page(keyword).url
        worked=True
        
    except wikipedia.exceptions.DisambiguationError as e:
        options = e.options
    except wikipedia.PageError as e:
        options = wikipedia.search(keyword)
        
        
    return options


## Define Similarity Function

nlp=spacy.load('en_core_web_sm')

def cosine_similarity(doc1,doc2): 
    """Calculate the similarity between original keyword and list of options output"""
    nlp1 = nlp(doc1)
    nlp2 = nlp(doc2)
    similarity = nlp1.similarity(nlp2)
    return similarity

## Define Option Evaluation Function

def select_option(options, keyword): #options is a list
    """ Create a function to evaluate option in options """
    
    if len(options)==1:
        option = options[0] # The only result will be used as searching query
    elif len(options)>1: 
        try: 
            temp=[]
            for option in options: 
                #score=cosine_similarity(wikipedia.summary(option,sentences=1),keyword)
                try:
                    score=cosine_similarity(wikipedia.summary(option),keyword)
                    #page = wikipedia.page(option)                    
                    
                    #score=cosine_similarity(page.content, keyword)
                    temp.append((option, score))
                except:
                    None
            option = temp #options[temp.index(max(temp))]
            
        except:
            option = None
    else:
        option = None
        print("***OPTION***", options)
    return option

## Define URL Query Function

def create_url(query):
    """Find URL for query"""
    page = wikipedia.page(query).url
    
    return page

## Get Query for Top 5

query_list = []
for word in keywords:
    options = get_options([word])
    if len(options) > 1:
        options = options + get_options([word+'_'])
        
    selected = select_option(options,summary + data_science)
    if type(selected) == list:
        selected.sort(key = lambda x: x[1], reverse=True)
        selected = selected[0][0]
    query_list.append(selected)
query_list

## Get URLs for Top 3

url_list = []
for query in query_list:
    sim = cosine_similarity(wikipedia.summary(query), summary + data_science)
    url = create_url(query)
    url_list.append((query, url, sim))
                    
url_list.sort(key = lambda x: x[2], reverse=True)
url_list = url_list[:3]
url_list = [(x, y) for x,y,z in url_list]
url_list 


## Apply to All Transcripts

def url_creation(lecture_keywords, lecture_summary):
    query_list = []
    url_list = []
    lecture_keywords_clean = clean_alt_list(lecture_keywords)
    for word in lecture_keywords_clean:     
        #gather all options
        options = get_options([word])
        if len(options) > 1:
            options = options + get_options([word+'_'])
        elif len(options) < 1:
            print(word, options) 
        #select the best option
        selected = select_option(options, lecture_summary + data_science)
        if type(selected) == list:
            selected.sort(key = lambda x: x[1], reverse=True)
            selected = selected[0][0]
        query_list.append(selected)
    res = []
    for val in query_list:
        if val != None :
            res.append(val)
    query_list = res
    for query in query_list:
        sim = cosine_similarity(wikipedia.summary(query), lecture_summary + data_science)
        url = create_url(query)
        url_list.append((query, url, sim))
                    
    url_list.sort(key = lambda x: x[2], reverse=True)
    if len(url_list) > 3:
        url_list = url_list[:3]
    url_list = [(x, y) for x,y,z in url_list]
    return url_list

df['url'] = df.apply(lambda x: url_creation(x.keywords, x.summary), axis=1)



df

## Save Dataset
We save the transcript dataset as a csv file for further analysis.

df.to_csv("./intermediate_data/transcripts_{}_summaries_keywords_urls".format(course))

## Next step
After you saved the dataset here, the information can then be used as part of a web app for usability.

---

**Authors:** [Wei Zhou](mailto:weiwzhou@umich.edu), [Nick Capaldini](mailto:nickcaps@umich.edu), University of Michigan, August 21, 2022

---
